supplier and customer module done

+++ app/Http/Controllers/Backend/UserController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers\Backend;
 
+use Image;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\File;
+
 
 class UserController extends Controller
 {
@@ -15,7 +18,7 @@ class UserController extends Controller
      */
     public function view()
     {
-        $data['allData'] = User::orderby('id','DESC')->get();
+        $data['allData'] = User::orderBy('id','desc')->get();
         return view('backend.pages.user.view-user',$data);
     }
 
@@ -39,6 +42,7 @@ class UserController extends Controller
     {
         $this->validate($request,[
             'user_type' => 'required',
+            'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
             'name' => 'required',
             'email' => 'required|email|unique:users,email',
             'password' => 'required|min:6|confirmed',      
@@ -47,11 +51,21 @@ class UserController extends Controller
     ]);
         $data = new User();
         $data->user_type = $request->user_type;
+      
+        if($request->image)
+        {
+          $image = $request->file('image') ;
+          $img = rand() . '.' .$image->getClientOriginalExtension();
+          $location = public_path('Backend/img/user/' . $img);
+          Image::make($image)->save($location);
+          $data->image = $img;
+        }
+
         $data->name = $request->name;
         $data->email = $request->email;
         $data->password = bcrypt($request->password);
         $data->save();
-        return redirect()->route('users.view')->with('success','Data Added Successfully');
+        return redirect()->route('users.view')->with('message','Data Added Successfully');
         
     }
 
@@ -94,23 +108,55 @@ class UserController extends Controller
             'email' => 'required|email',      
 
     ]);
-        // $data = User::find($id);
-        // $data->user_type = $request->user_type;
-        // $data->name = $request->name;
-        // $data->email = $request->email;
-        // $data->save();
         
-            $data = User::find($id);
+        
+        $data = User::find($id);
+        $data->user_type = $request->user_type;
+      
+        if(!is_null($request->image))
+        {
+            if( File::exists('Backend/img/user/' . $data->image) ){
+                File::delete('Backend/img/user/' . $data->image);
+            }
+          $image = $request->file('image') ;
+          $img = rand() . '.' .$image->getClientOriginalExtension();
+          $location = public_path('Backend/img/user/' . $img);
+          Image::make($image)->save($location);
+          $data->image = $img;
+        }
+
+        $data->name = $request->name;
+        $data->email = $request->email;
+        $data->password = bcrypt($request->password);
+        $data->save();
+        return redirect()->route('users.view')->with('info','Data Updated Successfully');
+        
+
+        //     if(!is_null($request->image))
+        // {
+        //     if( File::exists('Backend/img/user/' . $data->image) ){
+        //         File::delete('Backend/img/user/' . $data->image);
+        //     }
+        //     $image = $request->file('image') ;
+        //     $img = rand() . '.' .$image->getClientOriginalExtension();
+        //     $location = public_path('Backend/img/user/' . $img);
+        //     Image::make($image)->save($location);
+        //     $data->image = $img;
+
+        // }
+       
             
-            $data->update([
+        //     $data->update([
 
-                'user_type' => $request->input('user_type'),
-                'name' => $request->input('name'),
-                'email' => $request->input('email'),
+        //         'user_type' => $request->input('user_type'),
+        //         'name' => $request->input('name'),
+        //         'email' => $request->input('email'),
+        //         'image' =>$img,
                 
-            ]);
+                
+        //     ]);
 
-            return redirect()->route('users.view')->with('success','Data Updated Successfully');
+            //return redirect()->route('users.view')->with('info','Data Updated Successfully');
         
     }
 
@@ -123,7 +169,16 @@ class UserController extends Controller
     public function delete($id)
     {
         $deleteUser = User::find($id);
-        $deleteUser->delete();
-        return redirect()->route('users.view')->with('success','Data deleted successfully');
+        
+        if(!is_null($deleteUser))
+        {
+            if( File::exists('Backend/img/user/' . $deleteUser->image) ){
+                File::delete('Backend/img/user/' . $deleteUser->image);
+            }
+
+            $deleteUser->delete();
+        }
+       
+        return redirect()->route('users.view')->with('warning','Data deleted successfully');
     }
 }
